Find and Replacing Tool
A tool that allows you to find scene objects by name, type, tag and selection. It's useful for wanting to find and swap out items in bulk.

As an example;
You have a modular wall made of 20 pieces of the same object. It would look very monotone and uninteresting. You can use the tool and search by selection and select a few modular pieces and swap them out for another different piece. This allows you to quickly edit patterns without continuously dragging back and forth from the project folder and snapping them together.

It can customize the individual transform as well as keep the transform of the current object.